feat(#1357): add EnginePrefix.h + SynthEngine::getEngineParamPrefix() API
Move frozenPrefixForEngine() from PresetManager.h into a new minimal
header EnginePrefix.h to break the PresetManager dependency chain.
SynthEngine gains a non-virtual getEngineParamPrefix() that delegates
to frozenPrefixForEngine(getEngineId()) — RT-safe via static-local map.
No engine subclass needs to override this.
